Return usage of BMP085 pressure sensor. And use temperature compensation for it. #5 	modified:   flight/modules/Altitude/revolution/altitude.c 	modified:   flight/pios/common/pios_bmc050.c 	modified:   flight/pios/common/pios_bmp085_i2c.c 	modified:   flight/pios/inc/pios_bmp085_i2c.h 	modified:   flight/targets/boards/brainyfly/firmware/inc/pios_config.h

+++ flight/pios/inc/pios_bmp085_i2c.h
@@ -75,7 +75,7 @@ extern int32_t PIOS_BMP085_EOC;
 extern void PIOS_BMP085_Init(uint32_t i2c_id);
 extern void PIOS_BMP085_StartADC(ConversionTypeTypeDef Type);
 extern void PIOS_BMP085_ReadADC(void);
-extern int16_t PIOS_BMP085_GetTemperature(void);
+extern float PIOS_BMP085_GetTemperature(void);
 extern float PIOS_BMP085_GetPressure(void); // kpa
 extern bool PIOS_BMP085_Read(uint8_t address, uint8_t *buffer, uint8_t len);
 extern bool PIOS_BMP085_Write(uint8_t address, uint8_t buffer);
